+++ CHANGELOG.md
@@ -14,9 +14,12 @@ The **"Breaking Changes"** listed below are changes that have been made in the d
 
 ### Changed
 - Collapse the map list by default.
+- Collision view now has a transparency slider to help make it easier to view the underlying metatiles.
 
 ### Fixed
 - Fix bug where smart paths could bes auto-enabled, despite the checkbox being disabled.
+- Fix crash that could occur when changing the palette id in the tileset palette editor.
+- Fix crash that could occur when shrinking the number of metatiles in a tileset.
 
 ## [1.1.0] - 2018-12-27
 ### Breaking Changes